+++ packages/preact/src/head.ts
@@ -1,0 +1,8 @@
+export {
+  applyHead,
+  fromPageHead,
+  KEEL_HEAD_ATTR,
+  setTitle,
+  syncHead,
+  type HeadInput,
+} from "@kolektiv/keel"